backend/requirement.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==26.1.0
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
reportlab==4.2.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
